Add karma plugin to be able to parse HTML files looking for Angular inline templates
Templates defined as inline templates in an HTML file can't be handled by karma plugin ng-html2js as it treats only templates corresponding to a whole HTML file. However it is interesting to be able to parse some HTML files to find AngularJS inline templates and put them into cache that it can be used inside karma tests. This is what karma-ng-inline-templates-preprocessor is made for.
It could have been a totally independant project.